Add F6-F8 key support, pause/step controls, editor UI
- Added F6 (Play/Stop), F7 (Pause/Resume), F8 (Step) to KeyCode and InputSystem
- Implemented Game::togglePauseMode() and stepSimulation()
- Updated editor UI with Play, Stop, Pause/Resume, Step buttons and tooltips
- Handled F6-F8 in onInternalUpdate() for play state control
- Improved play mode transitions and viewport focus
- Changed getInternalKeyCode() return type to int

+++ DX3D/Include/DX3D/Input/InputSystem.h
@@ -25,7 +25,7 @@ namespace dx3d
 
 		void update();
 	private:
-		short getInternalKeyCode(const KeyCode& key);
+		int getInternalKeyCode(const KeyCode& key);
 		void centerCursor();
 	private:
 		std::array<bool, static_cast<std::size_t>(KeyCode::Count)> m_currentKeys{};
